v1.0.0 Release (#4)
Dropped support for .NET and PowerShell versions that are passed EOL:
  - .NET 6.0 and PowerShell 7.2
  - .NET 7.0 and PowerShell 7.3

Added support for .NET 9.0 and PowerShell 7.5

Corrected an Access Denied exception when trying to use GetDirectory or
GetFile and the executing user doesn't have access to a sibling
directory. #3

"Corrected" an issue in the unit tests where disposables weren't
releasing before the test cleanup routine was executing and failing,
causing the test to fail.